Refactoring the cve id reservation's general logic, priority and sequential unit tests to their own test files
