+++ test/unit-tests/cve-id/reserveCveId/cveIdReserveSequentialTest.js
@@ -1,0 +1,136 @@
+const express = require('express')
+const app = express()
+const chai = require('chai')
+const expect = chai.expect
+chai.use(require('chai-http'))
+
+// Body Parser Middleware
+app.use(express.json()) // Allows us to handle raw JSON data
+app.use(express.urlencoded({ extended: false })) // Allows us to handle url encoded data
+const middleware = require('../../../../src/middleware/middleware')
+app.use(middleware.createCtxAndReqUUID)
+
+const year20 = 2020
+const year21 = 2021
+const errors = require('../../../../src/controller/cve-id.controller/error')
+const error = new errors.CveIdControllerError()
+
+const cveIdFixtures = require('../mockObjects.cve-id')
+const cveIdController = require('../../../../src/controller/cve-id.controller/cve-id.controller')
+const cveIdParams = require('../../../../src/controller/cve-id.controller/cve-id.middleware')
+const repos = require('../../../../test-utils/repositories')
+
+describe('Testing the sequential reservation of GET /cve-id/:id endpoint in CveId Controller', () => {
+  context('Negative Tests', () => {
+    it('CveId Range document for year 2025 does not exist', (done) => {
+      app.route('/cve-id-reserve-2025-doesnt-exist-sequential')
+        .post((req, res, next) => {
+          const factory = {
+            getCveIdRepository: () => { return new repos.CveIdReserveSequentialYearDoesntExist() },
+            getCveIdRangeRepository: () => { return new repos.CveIdRangeReserveYearDoesntExist() },
+            getUserRepository: () => { return new repos.NullUserRepo() },
+            getOrgRepository: () => { return new repos.OrgReserveYear2025RangeDoesntExistSequential() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, cveIdParams.parsePostParams, cveIdController.CVEID_RESERVE)
+
+      chai.request(app)
+        .post(`/cve-id-reserve-2025-doesnt-exist-sequential?short_name=${cveIdFixtures.owningOrg.short_name}&cve_year=2025&amount=5&batch_type=sequential`)
+        .set(cveIdFixtures.secretariatHeader)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          const quotaHeader = (cveIdFixtures.owningOrg.policies.id_quota - 2).toString()
+          expect(res.header).to.have.property('cve-api-remaining-quota').and.to.equal(quotaHeader)
+
+          expect(res).to.have.status(403)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.cannotReserveForYear('2025')
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+
+    it('CveId Range document for year 2021 is full', (done) => {
+      app.route('/cve-id-reserve-2021-full-sequential')
+        .post((req, res, next) => {
+          const factory = {
+            getCveIdRepository: () => { return new repos.CveIdReserveSequentialIsFull() },
+            getCveIdRangeRepository: () => { return new repos.CveIdRangeReserveSequentialIsFull() },
+            getUserRepository: () => { return new repos.NullUserRepo() },
+            getOrgRepository: () => { return new repos.OrgReserveYear2025RangeDoesntExistSequential() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, cveIdParams.parsePostParams, cveIdController.CVEID_RESERVE)
+
+      chai.request(app)
+        .post(`/cve-id-reserve-2021-full-sequential?short_name=${cveIdFixtures.owningOrg.short_name}&cve_year=${year21}&amount=10&batch_type=sequential`)
+        .set(cveIdFixtures.secretariatHeader)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          const quotaHeader = (cveIdFixtures.owningOrg.policies.id_quota - 2).toString()
+          expect(res.header).to.have.property('cve-api-remaining-quota').and.to.equal(quotaHeader)
+
+          expect(res).to.have.status(403)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          const errObj = error.yearRangeFull(year21)
+          expect(res.body.error).to.equal(errObj.error)
+          expect(res.body.message).to.equal(errObj.message)
+          done()
+        })
+    })
+  })
+
+  context('Positive Tests', () => {
+    it('Cve IDs are reserved for year 2020', (done) => {
+      app.route('/cve-id-reserve-2020-sequential')
+        .post((req, res, next) => {
+          const factory = {
+            getCveIdRepository: () => { return new repos.CveIdReserveSequential() },
+            getCveIdRangeRepository: () => { return new repos.CveIdRangeReserveSequential() },
+            getUserRepository: () => { return new repos.UserReserveSequentialPriorityIsFull() },
+            getOrgRepository: () => { return new repos.OrgReserveSequentialPriorityIsFull() }
+          }
+          req.ctx.repositories = factory
+          next()
+        }, cveIdParams.parsePostParams, cveIdController.CVEID_RESERVE)
+
+      const top = 10
+      let counter = 1
+
+      chai.request(app)
+        .post(`/cve-id-reserve-2020-sequential?short_name=${cveIdFixtures.owningOrg.short_name}&cve_year=${year20}&amount=5&batch_type=sequential`)
+        .set(cveIdFixtures.owningOrgHeader)
+        .end((err, res) => {
+          if (err) {
+            done(err)
+          }
+
+          const quotaHeader = (cveIdFixtures.owningOrg.policies.id_quota - 7).toString() // already has two reserved cve ids
+          expect(res.header).to.have.property('cve-api-remaining-quota').and.to.equal(quotaHeader)
+
+          expect(res).to.have.status(200)
+          expect(res).to.have.property('body').and.to.be.a('object')
+          expect(res.body).to.have.property('cve_ids').and.to.be.a('array').and.to.have.lengthOf(5)
+          res.body.cve_ids.forEach(obj => {
+            const count = String(top + counter).padStart(4, '0')
+            expect(obj).to.have.property('cve_id').and.to.equal(`CVE-${year20}-${count}`) // check that cve ids are sequential
+            expect(obj).to.have.property('state').and.to.equal('RESERVED')
+            expect(obj).to.have.property('owning_cna').and.to.equal(cveIdFixtures.owningOrg.short_name)
+            expect(obj).to.have.nested.property('requested_by.cna').and.to.equal(cveIdFixtures.owningOrg.short_name)
+            expect(obj).to.have.nested.property('requested_by.user').and.to.equal(cveIdFixtures.owningOrgUser.username)
+            counter++
+          })
+          done()
+        })
+    })
+  })
+})
